+++ CHANGELOG.md
@@ -6,9 +6,18 @@ All issue numbers are relative to https://github.com/pywr/pywr/issues unless oth
 
 ### New features
 
-- Added flow weights to `AggregatedNode` (#603)
+- Allow solver to be defined by the environment variable `PYWR_SOLVER`. (#619)
+- Added flow weights to `AggregatedNode`. (#603)
 - Added additional labeling functionality to notebook graphing functions. (#612)
 - New and improved variable API for Parameters. (#601, #258)
+
+### Bug fixes
+
+- Fix a bug in `AnnualHarmonicSeriesParameter` related to updating the `amplitudes` and `phases` values with `set_double_variables` (#622)
+
+### Miscellaneous
+
+- Dropped support for Python 2 and <3.6. Pywr is no longer tested against Python versions earlier than 3.6. (#623)
 
 ## v0.5.1